requirements.txt
textual
openai
PyPDF2
python-docx
python-dotenv